refactor(phase-3-3): split step_group.py + export_monitor.py + tab builders
step_group.py (612) → 2 文件:
- step_group.py (313): StepGroup 主类
- _group_card.py (362): GroupCard 卡片

export_panel.py (675) → 减少 191 行:
- 提取 4 个 _create_*_tab 方法到 _tab_builders.py
- 现在 export_panel.py = 484 行 (从 675)

export_monitor.py (519) → 4 文件:
- export_monitor.py (45): 重导出
- _progress_widget.py (206): ExportProgressWidget
- _monitor_widget.py (237): ExportMonitorWidget
- _progress_dialog.py (165): ExportProgressDialog

修复:
- _monitor_widget.py 引用 ExportProgressWidget → 加 import
- _progress_dialog.py 引用 ExportMonitorWidget → 加 import
- ruff --fix 自动修 isort

验证:
- ruff check all green
- pytest 351 passed, 20 skipped
- 5/7 大文件拆分完成 (theme_optimizer/home_page/step_group/export_panel/export_monitor)
- 剩余: main_window.py(662) + monitor_panel.py(643) 需 PySide6 验证
