+++ src/scenefab/ui/main/pages/home_page.py
@@ -20,15 +20,14 @@ from PySide6.QtWidgets import (
     QWidget,
 )
 
-from ...theme.ds_tokens import Colors, FontSizes, FontWeights, Radii
-
 # ═══════════════════════════════════════════════════════════════════════
 # 玻璃态卡片
 # ═══════════════════════════════════════════════════════════════════════
-
-
 from scenefab.ui.main.pages._cards import GlassCard, ProjectCard, QuickCard, StatChip
 from scenefab.ui.main.pages._empty_state import EmptyStateCard
+
+from ...theme.ds_tokens import Colors, FontSizes, FontWeights, Radii
+
 
 class HomePage(QFrame):
     """首页"""
